+++ problems/matmul-swish-scaling/def.py
@@ -27,7 +27,7 @@ class matmul_swish_scaling(Problem):
         Returns:
             Result of scale * swish(A * B)
         """
-        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=torch.float32):
+        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=A.dtype):
             matmul_result = torch.matmul(A, B)
             swish_result = matmul_result * torch.sigmoid(matmul_result)
             scaled_result = scale * swish_result
