+++ problems/min-spanning-tree/def.py
@@ -25,7 +25,7 @@ class min_spanning_tree(Problem):
         Returns:
             Total weight of minimum spanning tree as a scalar tensor
         """
-        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=torch.float32):
+        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=adj_matrix.dtype):
             N = adj_matrix.size(0)
             device = adj_matrix.device
