+++ problems/scaled-dot-attention/def.py
@@ -27,7 +27,7 @@ class scaled_dot_attention(Problem):
         Returns:
             Result of scaled dot-product attention
         """
-        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=torch.float32):
+        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=query.dtype):
             # Use PyTorch's built-in scaled dot product attention
             return torch.nn.functional.scaled_dot_product_attention(
                 query, key, value, attn_mask=None, dropout_p=0.0, is_causal=False
